requirements.txt
# Pose LLM Identifier - Dependencies
# Updated as of November 5, 2025

# Core Deep Learning
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Transformers and Models
transformers>=4.35.0
timm>=0.9.12
huggingface-hub>=0.19.4

# Computer Vision
opencv-python>=4.8.1
albumentations>=1.3.1
Pillow>=10.1.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0

# Data Handling
pycocotools>=2.0.7
pandas>=2.1.0

# Configuration
pyyaml>=6.0.1
omegaconf>=2.3.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0

# Logging and Monitoring
tqdm>=4.66.0
wandb>=0.16.0
tensorboard>=2.15.0

# Utilities
scikit-learn>=1.3.0
einops>=0.7.0

# LLM Integration (Optional)
openai>=1.3.0
anthropic>=0.7.0

# For downloading models
gdown>=4.7.0
